mistral_ocr_processor/requirements.txt
mistralai>=1.0.0
pathlib>=1.0.1
unidecode>=1.3.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.1